+++ packages/antd/src/components/buttons/show/index.spec.tsx
@@ -1,13 +1,18 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
 
-import { fireEvent, render, TestWrapper, waitFor } from "@test";
+import { act, fireEvent, render, TestWrapper, waitFor } from "@test";
 import { ShowButton } from "./";
 
 describe("Show Button", () => {
     const show = jest.fn();
 
-    it("should render button successfuly", () => {
+    beforeAll(() => {
+        jest.spyOn(console, "warn").mockImplementation(jest.fn());
+        jest.useFakeTimers();
+    });
+
+    it("should render button successfuly", async () => {
         const { container, getByText } = render(
             <ShowButton onClick={() => show()} />,
             {
@@ -15,12 +20,16 @@ describe("Show Button", () => {
             },
         );
 
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
+        });
+
         expect(container).toBeTruthy();
 
         getByText("Show");
     });
 
-    it("should render text by children", () => {
+    it("should render text by children", async () => {
         const { container, getByText } = render(
             <ShowButton>refine</ShowButton>,
             {
@@ -28,14 +37,22 @@ describe("Show Button", () => {
             },
         );
 
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
+        });
+
         expect(container).toBeTruthy();
 
         getByText("refine");
     });
 
-    it("should render without text show only icon", () => {
+    it("should render without text show only icon", async () => {
         const { container, queryByText } = render(<ShowButton hideText />, {
             wrapper: TestWrapper({}),
+        });
+
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
         });
 
         expect(container).toBeTruthy();
@@ -52,11 +69,13 @@ describe("Show Button", () => {
             }),
         });
 
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
+        });
+
         expect(container).toBeTruthy();
 
-        await waitFor(() =>
-            expect(getByText("Show").closest("button")).toBeDisabled(),
-        );
+        expect(getByText("Show").closest("button")).toBeDisabled();
     });
 
     it("should be disabled when recordId not allowed", async () => {
@@ -76,11 +95,13 @@ describe("Show Button", () => {
             },
         );
 
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
+        });
+
         expect(container).toBeTruthy();
 
-        await waitFor(() =>
-            expect(getByText("Show").closest("button")).toBeDisabled(),
-        );
+        expect(getByText("Show").closest("button")).toBeDisabled();
     });
 
     it("should skip access control", async () => {
@@ -95,11 +116,13 @@ describe("Show Button", () => {
             },
         );
 
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
+        });
+
         expect(container).toBeTruthy();
 
-        await waitFor(() =>
-            expect(getByText("Show").closest("button")).not.toBeDisabled(),
-        );
+        expect(getByText("Show").closest("button")).not.toBeDisabled();
     });
 
     it("should successfully return disabled button custom title", async () => {
@@ -115,11 +138,13 @@ describe("Show Button", () => {
             }),
         });
 
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
+        });
+
         expect(container).toBeTruthy();
 
-        await waitFor(() =>
-            expect(getByText("Show").closest("button")).not.toBeDisabled(),
-        );
+        expect(getByText("Show").closest("button")).toBeDisabled();
         await waitFor(() =>
             expect(
                 getByText("Show").closest("button")?.getAttribute("title"),
@@ -127,17 +152,23 @@ describe("Show Button", () => {
         );
     });
 
-    it("should render called function successfully if click the button", () => {
+    it("should render called function successfully if click the button", async () => {
         const { getByText } = render(<ShowButton onClick={() => show()} />, {
             wrapper: TestWrapper({}),
         });
 
-        fireEvent.click(getByText("Show"));
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
+        });
+
+        await act(async () => {
+            fireEvent.click(getByText("Show"));
+        });
 
         expect(show).toHaveBeenCalledTimes(1);
     });
 
-    it("should create page redirect show route called function successfully if click the button", () => {
+    it("should create page redirect show route called function successfully if click the button", async () => {
         const { getByText } = render(
             <Routes>
                 <Route
@@ -153,12 +184,18 @@ describe("Show Button", () => {
             },
         );
 
-        fireEvent.click(getByText("Show"));
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
+        });
+
+        await act(async () => {
+            fireEvent.click(getByText("Show"));
+        });
 
         expect(window.location.pathname).toBe("/posts/show/1");
     });
 
-    it("should edit page redirect show route called function successfully if click the button", () => {
+    it("should edit page redirect show route called function successfully if click the button", async () => {
         const { getByText } = render(
             <Routes>
                 <Route path="/:resource/:action/:id" element={<ShowButton />} />
@@ -171,12 +208,18 @@ describe("Show Button", () => {
             },
         );
 
-        fireEvent.click(getByText("Show"));
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
+        });
+
+        await act(async () => {
+            fireEvent.click(getByText("Show"));
+        });
 
         expect(window.location.pathname).toBe("/posts/show/1");
     });
 
-    it("should custom resource and recordItemId redirect show route called function successfully if click the button", () => {
+    it("should custom resource and recordItemId redirect show route called function successfully if click the button", async () => {
         const { getByText } = render(
             <Routes>
                 <Route
@@ -197,12 +240,18 @@ describe("Show Button", () => {
             },
         );
 
-        fireEvent.click(getByText("Show"));
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
+        });
+
+        await act(async () => {
+            fireEvent.click(getByText("Show"));
+        });
 
         expect(window.location.pathname).toBe("/categories/show/1");
     });
 
-    it("should redirect with custom route called function successfully if click the button", () => {
+    it("should redirect with custom route called function successfully if click the button", async () => {
         const { getByText } = render(
             <Routes>
                 <Route
@@ -229,7 +278,13 @@ describe("Show Button", () => {
             },
         );
 
-        fireEvent.click(getByText("Show"));
+        await act(async () => {
+            jest.advanceTimersToNextTimer(1);
+        });
+
+        await act(async () => {
+            fireEvent.click(getByText("Show"));
+        });
 
         expect(window.location.pathname).toBe("/custom-route-posts/show/1");
     });